Change data type associated with money from double to BigDecimal

+++ src/main/java/com/auction/models/Admin.java
@@ -2,6 +2,8 @@ package com.auction.models;
 
 import com.auction.server.factory.UserRole;
 
+import java.math.BigDecimal;
+
 public class Admin extends User {
     private static final long serialVersionUID = 1L;
 
@@ -11,7 +13,7 @@ public class Admin extends User {
     }
 
     // Create an Admin instance from an existed Admin in DB
-    public Admin(String id, String username, String password, double balance) {
+    public Admin(String id, String username, String password, BigDecimal balance) {
         super(id, UserRole.ADMIN, username, password, balance);
     }
 }
