+++ src/main/java/com/auction/models/Bidder.java
@@ -2,6 +2,8 @@ package com.auction.models;
 
 import com.auction.server.factory.UserRole;
 
+import java.math.BigDecimal;
+
 public class Bidder extends User {
     private static final long serialVersionUID = 1L;
 
@@ -16,7 +18,7 @@ public class Bidder extends User {
     }
 
     // Create a Bidder instance from an existed Bidder in DB
-    public Bidder(String id, String username, String password, double balance) {
+    public Bidder(String id, String username, String password, BigDecimal balance) {
         super(id, UserRole.BIDDER, username, password, balance);
     }
 }
